Allow for irregular pull widths
Assuming the axis has an `edges` attribute, I assume it should also have a
`widths` attribute, so I thought we can just use that for the pull bar widths
instead of calculating our own assuming a regular spacing (which might not be
true).

For the patch-width, we just use `right_edge - left_edge`, instead of what was
done before, which dividing this difference by the number of pulls and then
multiplying them again with the same number.