Fix the tests by using re-seeding the global bit-generator
Most tests use `np.random.seed(42)` to re-seed the global RNG. Initially I didn't
follow this in my tests, because according to its documentation it is not
best-practice and a legacy function.

> The best practice is to not reseed a BitGenerator, rather to recreate a new
> one. This method is here for legacy reasons.

Instead, I created a custom RNG with a seed and used that to generate my random
numbers. However, the tests failed and I noticed that the random seed not only affects the numbers to
fill the histogram with, but also the fit, which I can't affect via a custom RNG,
so I have to re-seed the global RNG.

Therefore, I fall back to using `np.random.normal`.