feat: add `fill_flattened` support (#474)
This PR draft is based upon
https://gist.github.com/agoose77/f755d96061a36795dda592276d63f506, which
prototyped what `fill_flattened` might look like in Hist. The purpose of
this feature is to make histogramming with different-structure arrays
more convenient.

This implementation is fairy opinionated; it should be possible, for
example to support pass in axis arrays by destructuring obejcts _and_
keyword argument. However, in the pursuit of simplicity, I've made these
mutually exclusive; either a single structured object should be used, or
the `*args` and `**kwargs` are used as expected to map onto axes.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>