fix: `fill_flattened` broadcasting of non user kwargs (#612)
When using`fill_flattened` with multiple args and either the `weight` or
`sample` kwargs, it tries to broadcast the keys of `non_user_kwargs`
instead of the values, leading to kwargs trying to fill the histogram
with the stringified kwarg name instead of its value (such as the
`sample` kwarg attempting to fill the array with `["sample", "sample",
"sample", "sample"]`). This PR changes that behavior to insert the
values instead.